improvements in ValueGeneratorComponent
